pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opensage"
version = "0.1.0"
description = "OpenSage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pip",
    "google-adk<=1.25.1",
    "pandas==2.2.3",
    "neomodel==5.4.3",
    "docker==7.1.0",
    "paramiko>=3.0.0",
    "python-dotenv==1.1.0",
    "tree-sitter-languages==1.10.2",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "litellm>=1.74.7",
    "msgspec>=0.19.0",
    "websockets>=15.0.1",
    "networkx>=3.4.2",
    "toml>=0.10.2",
    "fire>=0.6.0",
    "jsonpickle>=3.0.0",
    "datasets>=2.0.0",
    "dacite>=1.8.0",
    "cxxfilt>=0.3.0",
    "click>=8.1.8,<9.0.0",
    "uvicorn>=0.34.0,<1.0.0",
    "python-magic>=0.4.27",
    "opensandbox>=0.1.5",
    "agentdocker-lite @ git+https://github.com/rucnyz/agentdocker-lite.git@main",
]

[project.optional-dependencies]
langfuse = [
    "langfuse==3.11.1",
    "openinference-instrumentation-google-adk==0.1.8",
    "opentelemetry-api==1.37.0",
    "opentelemetry-sdk==1.37.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.37.0",
    "opentelemetry-exporter-otlp-proto-http==1.37.0"
]
cybergym = [
    "cybergym[dev,server] @ git+https://github.com/sunblaze-ucb/cybergym",
]
full = [
    "opensage[langfuse,cybergym]",
]

[project.scripts]
opensage = "opensage.cli.opensage_cli:main"

[dependency-groups]
dev = [
    "pre-commit",
    "ipython",
    "pytest",
    "pytest-asyncio",
    "pytest-timeout",
    "ruff",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-autorefs>=1.4.0",
    "mkdocs-material>=9.0.0",
    "mkdocs-minify-plugin>=0.7.0",
    "mkdocstrings[python]>=0.30.0",
    "pymdown-extensions>=10.16.0",
]

[tool.uv]
override-dependencies = []
no-build-isolation-package = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = [
    "src/opensage",
    "examples",
]

[tool.hatch.build.targets.wheel]
packages = ["src/opensage"]

[tool.ruff]
extend-exclude = ["third_party"]

[tool.ruff.lint]
select = [
    "I", # isort
    "F", # pyflakes
]
ignore = ["F401"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["opensage", "tests"]

[tool.hatch.build.targets.wheel.force-include]
"examples" = "opensage/examples"
"src/opensage/cli/assets/opensage_favicon.ico" = "opensage/cli/assets/opensage_favicon.ico"
"src/opensage/cli/vendor/adk_browser" = "opensage/cli/vendor/adk_browser"
"src/opensage/plugins/default/claude_code_hooks" = "opensage/plugins/default/claude_code_hooks"

[tool.pytest.ini_options]
log_format = "%(asctime)s [%(name)s] [%(levelname)s] %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_level = "WARNING"
log_cli = false
timeout = 300
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
norecursedirs = ["tests/rl"]
markers = [
    "native_backend: tests for the native sandbox backend",
    "k8s_backend: tests for the Kubernetes sandbox backend",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
